BREAKING CHANGE: rename `Source` class to `PushSource` (#110)
https://coveord.atlassian.net/browse/CDX-876
