+++ samples/createSource.ts
@@ -1,7 +1,7 @@
-import {Source, SourceVisibility} from '@coveo/push-api-client';
+import {PushSource, SourceVisibility} from '@coveo/push-api-client';
 
 async function main() {
-  const source = new Source('my_api_key', 'my_coveo_organization_id');
+  const source = new PushSource('my_api_key', 'my_coveo_organization_id');
 
   const response = await source.create(
     'the_name_of_my_source',
