+++ samples/pushBatchOfDocuments.ts
@@ -1,10 +1,10 @@
 import {
-  Source,
+  PushSource,
   BatchUpdateDocuments,
   DocumentBuilder,
 } from '@coveo/push-api-client';
 async function main() {
-  const source = new Source('my_api_key', 'my_coveo_organization_id');
+  const source = new PushSource('my_api_key', 'my_coveo_organization_id');
   source.setSourceStatus('my_source_id', 'REFRESH');
 
   const myBatchOfDocuments: BatchUpdateDocuments = {
