+++ src/localtest.ts
@@ -6,10 +6,10 @@ const API_KEY = process.env.API_KEY as string;
 const ORG_ID = process.env.ORG_ID as string;
 const SOURCE_ID = process.env.SOURCE_ID as string;
 
-import {BatchUpdateDocuments, Source} from './source';
+import {BatchUpdateDocuments, PushSource} from './source/push';
 
 async function main() {
-  const source = new Source(API_KEY, ORG_ID);
+  const source = new PushSource(API_KEY, ORG_ID);
   await source.setSourceStatus(SOURCE_ID, 'REFRESH');
   const docBuilder = new DocumentBuilder(
     'https://perdu.com',
