+++ src/source/push.ts
@@ -13,24 +13,24 @@ import {
 } from '@coveord/platform-client';
 export {SourceVisibility} from '@coveord/platform-client';
 import axios, {AxiosRequestConfig, AxiosResponse} from 'axios';
-import {DocumentBuilder} from './documentBuilder';
+import {DocumentBuilder} from '../documentBuilder';
 import dayjs = require('dayjs');
 import {URL} from 'url';
-import {consumeGenerator} from './help/generator';
-import {parseAndGetDocumentBuilderFromJSONDocument} from './validation/parseFile';
+import {consumeGenerator} from '../help/generator';
+import {parseAndGetDocumentBuilderFromJSONDocument} from '../validation/parseFile';
 import {basename} from 'path';
-import {getAllJsonFilesFromEntries} from './help/file';
+import {getAllJsonFilesFromEntries} from '../help/file';
 import {
   castEnvironmentToPlatformClient,
   DEFAULT_ENVIRONMENT,
   DEFAULT_REGION,
   platformUrl,
   PlatformUrlOptions,
-} from './environment';
-import {FieldAnalyser} from './fieldAnalyser/fieldAnalyser';
-import {FieldTypeInconsistencyError} from './errors/fieldErrors';
-import {createFields} from './fieldAnalyser/fieldUtils';
-import {SecurityIdentity} from './source/securityIdenty';
+} from '../environment';
+import {FieldAnalyser} from '../fieldAnalyser/fieldAnalyser';
+import {FieldTypeInconsistencyError} from '../errors/fieldErrors';
+import {createFields} from '../fieldAnalyser/fieldUtils';
+import {SecurityIdentity} from './securityIdenty';
 
 export type SourceStatus = 'REBUILD' | 'REFRESH' | 'INCREMENTAL' | 'IDLE';
 
@@ -85,7 +85,7 @@ interface FileContainerResponse {
  *
  * Allows you to create a new push source, manage security identities and documents in a Coveo organization.
  */
-export class Source {
+export class PushSource {
   private platformClient: PlatformClient;
   private options: Required<PlatformUrlOptions>;
   private static defaultOptions: Required<PlatformUrlOptions> = {
@@ -103,7 +103,7 @@ export class Source {
     private organizationid: string,
     options?: PlatformUrlOptions
   ) {
-    this.options = {...Source.defaultOptions, ...options};
+    this.options = {...PushSource.defaultOptions, ...options};
     this.platformClient = new PlatformClient({
       accessToken: this.apikey,
       environment: castEnvironmentToPlatformClient(this.options.environment),
@@ -136,10 +136,6 @@ export class Source {
     securityProviderId: string,
     securityIdentity: SecurityIdentityModel
   ) {
-    console.log('This method has been deprecated');
-    console.log(
-      'Use `source.identity.createOrUpdateSecurityIdentity()` instead'
-    );
     return this.identity.createSecurityIdentity(
       securityProviderId,
       securityIdentity
@@ -155,10 +151,6 @@ export class Source {
     securityProviderId: string,
     securityIdentityAlias: SecurityIdentityAliasModel
   ) {
-    console.log('This method has been deprecated');
-    console.log(
-      'Use `source.identity.createOrUpdateSecurityIdentityAlias()` instead'
-    );
     return this.identity.createOrUpdateSecurityIdentityAlias(
       securityProviderId,
       securityIdentityAlias
@@ -174,8 +166,6 @@ export class Source {
     securityProviderId: string,
     securityIdentityToDelete: SecurityIdentityDelete
   ) {
-    console.log('This method has been deprecated');
-    console.log('Use `source.identity.deleteSecurityIdentity()` instead');
     return this.identity.deleteSecurityIdentity(
       securityProviderId,
       securityIdentityToDelete
@@ -191,8 +181,6 @@ export class Source {
     securityProviderId: string,
     batchDelete: SecurityIdentityDeleteOptions
   ) {
-    console.log('This method has been deprecated');
-    console.log('Use `source.identity.deleteOldSecurityIdentities()` instead');
     return this.identity.deleteOldSecurityIdentities(
       securityProviderId,
       batchDelete
@@ -208,8 +196,6 @@ export class Source {
     securityProviderId: string,
     batchConfig: SecurityIdentityBatchConfig
   ) {
-    console.log('This method has been deprecated');
-    console.log('Use `source.identity.manageSecurityIdentities()` instead');
     return this.identity.manageSecurityIdentities(
       securityProviderId,
       batchConfig
@@ -486,7 +472,7 @@ export class Source {
           JSON.stringify(docBuilder.marshal())
         );
 
-        if (accumulator.size + sizeOfDoc >= Source.maxContentLength) {
+        if (accumulator.size + sizeOfDoc >= PushSource.maxContentLength) {
           const chunks = accumulator.chunks;
           if (chunks.length > 0) {
             batchesToUpload.push(() =>
