+++ src/source/source.ts
@@ -1,0 +1,6 @@
+import {PushSource} from '..';
+
+/**
+ * @deprecated This class has been replaced by {@link PushSource}
+ */
+export const Source = PushSource;
